requirements.txt
openai-whisper
g4f
ffmpeg-python
gradio
PyExecJS
# -i https://pypi.douban.com/simple
# pip uninstall charset-normalizer
# conda install -c conda-forge charset-normalizer